ak-input: Input as part of state object. Project structure improvements.
